reorganise Boost serialization headers
